refactor(Modal): libs로 분리
